+++ src/main/java/no/nav/tag/tiltaksgjennomforing/avtale/EndreStillingsbeskrivelse.java
@@ -10,6 +10,6 @@ public class EndreStillingsbeskrivelse {
     String arbeidsoppgaver;
     Integer stillingStyrk08;
     Integer stillingKonseptId;
-    Integer stillingprosent;
+    Double stillingprosent;
     Double antallDagerPerUke;
 }
